requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-json-widget==1.1.1
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
pkg_resources==0.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
